Download and return a data.table object GSOD weather station data inventories

The NCEI maintains a document, https://www1.ncdc.noaa.gov/pub/data/noaa/isd-inventory.txt, which lists the number of weather observations by station-year-month from the beginning of the stations' records. This function retrieves that document and prints an information header displaying the last update time with a data frame of the inventory information for each station-year-month.

get_inventory()

Value

A data.table object of station inventories

Note

The GSOD data, which are downloaded and manipulated by GSODR, stipulate that the following notice should be given. “The following data and products may have conditions placed on their international commercial use. They can be used within the U.S. or for non- commercial international activities without restriction. The non-U.S. data cannot be redistributed for commercial purposes. Re-distribution of these data by others must provide this same notification.”

Author

Adam H. Sparks, adamhsparks@gmail.com

Examples

# \donttest{
inventory <- get_inventory()
inventory

#>   *** FEDERAL CLIMATE COMPLEX INTEGRATED SURFACE DATA INVENTORY ***  
#>    This inventory provides the number of weather observations by  
#>    STATION-YEAR-MONTH for beginning of record through March 2021   
#>                STNID NAME LAT LON CTRY STATE BEGIN END COUNTRY_NAME ISO2C ISO3C
#>      1: 007018-99999 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#>      2: 007018-99999 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#>      3: 007026-99999 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#>      4: 007026-99999 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#>      5: 007026-99999 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#>     ---                                                                        
#> 656858:   A51256-451 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#> 656859:   A51256-451 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#> 656860:   A51256-451 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#> 656861:   A51256-451 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#> 656862:   A51256-451 <NA>  NA  NA <NA>  <NA>    NA  NA         <NA>  <NA>  <NA>
#>         YEAR  JAN  FEB  MAR  APR  MAY  JUN  JUL  AUG  SEP  OCT  NOV  DEC
#>      1: 2011    0    0 2104 2797 2543 2614  382    0    0    0    0    0
#>      2: 2013    0    0    0    0    0    0  710    0    0    0    0    0
#>      3: 2012    0    0    0    0    0    0  367    0    0    0    0    7
#>      4: 2014    0    0    0    0    0    0  180    0    4    0  552    0
#>      5: 2016    0    0    0    0    0  794    0    0    0    0    0    0
#>     ---                                                                 
#> 656858: 2017 2192 1883 2204 1910 2145 2113 2218 2204 2082 2192 2103 2174
#> 656859: 2018 2192 1887 2194 2113 2151 2095 2202 2197 1816 2195 2063 2178
#> 656860: 2019 2188 2000 2143 2105 2187 2124 2184 2138 2077 1872 1508 2159
#> 656861: 2020 2165 1455 2144 2125 2199 2123 2112 2192 2083 2079 2074 2187
#> 656862: 2021 2085 1992 1031    0    0    0    0    0    0    0    0    0

# }