Download the Latest Station List From the NCEI Server

This function downloads the latest station list from the NCEI FTP server. This list includes metadata for all stations including unique identifiers, country, state (if in US), latitude, longitude, elevation and when weather observations begin and end. Stations with invalid latitude and longitude values will not be included.

get_station_list()

Value

data.table object of station metadata.

Examples

## Not run: ------------------------------------
# GSOD_stations <- get_station_list()
## ---------------------------------------------