GSODR: Global Surface Summary of the Day (GSOD) Weather Data from R
Source: R/GSODR-package.R

Provides automated downloading, parsing, cleaning, unit conversion and formatting of Global Surface Summary of the Day (GSOD) weather data from the from the USA's National Oceanic and Atmospheric Administration's (NOAA) National Centre for Environmental Information (NCEI) for use in R. All units are converted from from United States Customary System (USCS) units to International System of Units (SI). Stations may be individually checked for number of missing days defined by the user, where stations with too many missing observations are omitted. Only stations with valid reported latitude and longitude values are permitted in the final data. Additional useful elements, saturation vapour pressure (es), actual vapour pressure (ea) and relative humidity are calculated from the original data and included in the final data set. The resulting data include station identification information, state, country, latitude, longitude, elevation, weather observations and associated flags. Additional data are included with this R package: a list of elevation values for stations between -60 and 60 degrees latitude derived from the Shuttle Radar Topography Measuring Mission (SRTM). For information on the GSOD data from NCEI, please see the GSOD readme.txt file available from, http://www1.ncdc.noaa.gov/pub/data/gsod/readme.txt.

Source

https://data.noaa.gov/dataset/dataset/global-surface-summary-of-the-day-gsod/

Details

For climate data that have been formatted specifically for use with the GSODR package, please see the GSODRdata package (Sparks et al.) available on GitHub: https://adamhsparks.github.io/GSODRdata/. Four data frames of climate data are provided from various sources for GSOD station locations.

References

Karger, D. N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza, R. W., et al. (2016) Climatologies at high resolution for the Earth land surface areas. arXiv preprint arXiv:1607.00217

New, M., Lister, D., Hulme, M., Makin, I., (2002) A high-resolution data set of surface climate over global land areas. Climate Research 21:1--25

Sparks A., Hengl T., Nelson A. (2017) GSODRdata: Extra Climate Data for the GSODR Package. R package version 0.1.0, https://adamhsparks.github.io/GSODRdata/index.html.

Wilson A. M., Jetz W. (2016) Remotely Sensed High-Resolution Global Cloud Dynamics for Predicting Ecosystem and Biodiversity Distributions. PLoS Biol 14(3): e1002415. doi:10.1371/journal. pbio.1002415

See also

GSODR functions:

get_GSOD Download, Clean, Reformat Generate New Elements and Return a Tidy Data Frame of GSOD Weather Data

reformat_GSOD Clean, Reformat Generate New Elements and Return a Tidy Data Frame of GSOD Weather Data from Local Disk

nearest_stations Find Nearest GSOD Stations to Specified a Latitude and Longitude

update_station_list Download the Latest Station List Information and Update GSODR's Internal Database

get_inventory Download and return a tidy data frame of GSOD weather station data inventories

Useful links:

Static documentation at https://ropensci.github.io/GSODR/

Development repository at https://github.com/ropensci/GSODR

Report bugs at https://github.com/ropensci/GSODR/issues