Download and return a data.table object of GSOD weather data

This function automates downloading, cleaning, reformatting of data from the Global Surface Summary of the Day (GSOD) data provided by the US National Centers for Environmental Information (NCEI), Three additional useful elements: saturation vapour pressure (es), actual vapour pressure (ea) and relative humidity (RH) are calculated and returned in the final data frame using the improved August-Roche-Magnus approximation (Alduchov and Eskridge 1996).

get_GSOD(
  years,
  station = NULL,
  country = NULL,
  max_missing = NULL,
  agroclimatology = FALSE
)

Arguments

years
Year(s) of weather data to download.

station
Optional. Specify a station or multiple stations for which to retrieve, check and clean weather data using STATION. The NCEI reports years for which the data are available. This function checks against these years. However, not all cases are properly documented and in some cases files may not exist for download even though it is indicated that data was recorded for the station for a particular year. If a station is specified that does not have an existing file on the server, this function will silently fail and move on to existing files for download and cleaning.

country
Optional. Specify a country for which to retrieve weather data; full name, 2 or 3 letter ISO or 2 letter FIPS codes can be used. All stations within the specified country will be returned.

max_missing
Optional. The maximum number of days allowed to be missing from a station's data before it is excluded from final file output.

agroclimatology
Optional. Logical. Only clean data for stations between latitudes 60 and -60 for agroclimatology work, defaults to `FALSE`. Set to `TRUE` to include only stations within the confines of these latitudes.

Value

A data frame as a data.table object of GSOD weather data.

Details

All units are converted to International System of Units (SI), e.g., Fahrenheit to Celsius and inches to millimetres.

Data summarise each year by station, which include vapour pressure and relative humidity elements calculated from existing data in GSOD.

All missing values in resulting files are represented as NA regardless of which field they occur in.

For a complete list of the fields and description of the contents and units, please refer to Appendix 1 in the GSODR vignette, vignette("GSODR", package = "GSODR").

For more information see the description of the data provided by NCEI, https://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt.

Note

GSODR attempts to validate year and station combination requests, however, in certain cases the start and end date may encompass years where no data is available. In these cases no data will be returned. It is suggested that the user check the latest data availability for the station(s) desired using get_inventory as this list is frequently updated by the NCEI and is not shipped with GSODR.

While GSODR does not distribute GSOD weather data, users of the data should note the conditions that the U.S. NCEI places upon the GSOD data. “The following data and products may have conditions placed on their international commercial use. They can be used within the U.S. or for non-commercial international activities without restriction. The non-U.S. data cannot be redistributed for commercial purposes. Re-distribution of these data by others must provide this same notification.”

References

Alduchov, O.A. and Eskridge, R.E., 1996. Improved Magnus form approximation of saturation vapor pressure. Journal of Applied Meteorology and Climatology, 35(4), pp.601-609. DOI: <10.1175

See also

Author

Adam H. Sparks, adamhsparks@gmail.com

Examples

# \donttest{
# Download weather station for Toowoomba, Queensland for 2010
tbar <- get_GSOD(years = 2010, station = "955510-99999")

tbar

#>             STNID              NAME CTRY STATE LATITUDE LONGITUDE ELEVATION
#>   1: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#>   2: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#>   3: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#>   4: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#>   5: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#>  ---                                                                       
#> 361: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#> 362: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#> 363: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#> 364: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#> 365: 955510-99999 TOOWOOMBA AIRPORT   AS         -27.55   151.917       642
#>         BEGIN      END   YEARMODA YEAR MONTH DAY YDAY TEMP TEMP_ATTRIBUTES DEWP
#>   1: 19980301 20210116 2010-01-01 2010     1   1    1 21.2               8 17.9
#>   2: 19980301 20210116 2010-01-02 2010     1   2    2 23.2               8 19.4
#>   3: 19980301 20210116 2010-01-03 2010     1   3    3 21.4               8 18.9
#>   4: 19980301 20210116 2010-01-04 2010     1   4    4 18.9               8 16.4
#>   5: 19980301 20210116 2010-01-05 2010     1   5    5 20.5               8 16.4
#>  ---                                                                           
#> 361: 19980301 20210116 2010-12-27 2010    12  27  361 18.9               8 17.3
#> 362: 19980301 20210116 2010-12-28 2010    12  28  362 17.7               8 14.7
#> 363: 19980301 20210116 2010-12-29 2010    12  29  363 19.8               8 15.4
#> 364: 19980301 20210116 2010-12-30 2010    12  30  364 21.4               8 16.1
#> 365: 19980301 20210116 2010-12-31 2010    12  31  365 21.1               8 15.7
#>      DEWP_ATTRIBUTES    SLP SLP_ATTRIBUTES   STP STP_ATTRIBUTES VISIB
#>   1:               8 1013.4              8 942.0              8    NA
#>   2:               8 1010.5              8 939.3              8    NA
#>   3:               8 1012.3              8 940.9              8  14.3
#>   4:               8 1015.7              8 944.1              8  23.3
#>   5:               8 1015.5              8 944.0              8    NA
#>  ---                                                                 
#> 361:               8 1005.8              8 934.8              8  25.3
#> 362:               8 1013.2              8 941.6              8    NA
#> 363:               8 1014.4              8 942.7              8    NA
#> 364:               8 1014.6              8 942.9              8    NA
#> 365:               8 1013.9              8 942.3              8    NA
#>      VISIB_ATTRIBUTES WDSP WDSP_ATTRIBUTES MXSPD GUST  MAX MAX_ATTRIBUTES  MIN
#>   1:                0  4.3               8   6.7   NA 25.8           <NA> 17.8
#>   2:                0  3.7               8   5.1   NA 26.5           <NA> 19.1
#>   3:                6  7.6               8  10.3   NA 28.7           <NA> 19.3
#>   4:                4  8.7               8  10.3   NA 24.1           <NA> 16.9
#>   5:                0  7.5               8  10.8   NA 24.6           <NA> 16.7
#>  ---                                                                          
#> 361:                5  7.2               8  11.8   NA 22.1           <NA> 17.9
#> 362:                0 11.2               8  14.4   NA 21.5           <NA> 15.4
#> 363:                0  6.6               8   9.3   NA 24.0           <NA> 15.0
#> 364:                0  7.2               8   9.3   NA 26.5           <NA> 15.0
#> 365:                0  7.5               8   9.3   NA 26.5           <NA> 16.5
#>      MIN_ATTRIBUTES PRCP PRCP_ATTRIBUTES SNDP I_FOG I_RAIN_DRIZZLE I_SNOW_ICE
#>   1:           <NA>  1.5               G   NA    NA             NA         NA
#>   2:           <NA>  0.3               G   NA    NA             NA         NA
#>   3:              * 19.8               G   NA    NA             NA         NA
#>   4:              *  1.0               G   NA    NA             NA         NA
#>   5:           <NA>  0.3               G   NA    NA             NA         NA
#>  ---                                                                         
#> 361:              * 42.9               G   NA    NA             NA         NA
#> 362:              *  0.3               G   NA    NA             NA         NA
#> 363:           <NA>  0.0               G   NA    NA             NA         NA
#> 364:           <NA>  0.0               G   NA    NA             NA         NA
#> 365:           <NA>  0.0               G   NA    NA             NA         NA
#>      I_HAIL I_THUNDER I_TORNADO_FUNNEL  EA  ES   RH
#>   1:     NA        NA               NA 2.0 2.5 81.5
#>   2:     NA        NA               NA 2.2 2.8 79.2
#>   3:     NA        NA               NA 2.2 2.5 85.7
#>   4:     NA        NA               NA 1.9 2.2 85.4
#>   5:     NA        NA               NA 1.9 2.4 77.3
#>  ---                                               
#> 361:     NA        NA               NA 2.0 2.2 90.4
#> 362:     NA        NA               NA 1.7 2.0 82.6
#> 363:     NA        NA               NA 1.7 2.3 75.8
#> 364:     NA        NA               NA 1.8 2.5 71.8
#> 365:     NA        NA               NA 1.8 2.5 71.3

# Download data for Australia from 2010 to 2011
AUS <- get_GSOD(years = 2010:2011, country = "Australia")

AUS

#>                STNID                      NAME CTRY STATE LATITUDE LONGITUDE
#>      1: 941030-99999         BROWSE ISLAND AWS   AS        -14.117   123.533
#>      2: 941030-99999         BROWSE ISLAND AWS   AS        -14.117   123.533
#>      3: 941030-99999         BROWSE ISLAND AWS   AS        -14.117   123.533
#>      4: 941030-99999         BROWSE ISLAND AWS   AS        -14.117   123.533
#>      5: 941030-99999         BROWSE ISLAND AWS   AS        -14.117   123.533
#>     ---                                                                     
#> 185660: 959970-99999 HEARD ISLAND (ATLAS COVE)   AS        -53.017    73.400
#> 185661: 959970-99999 HEARD ISLAND (ATLAS COVE)   AS        -53.017    73.400
#> 185662: 959970-99999 HEARD ISLAND (ATLAS COVE)   AS        -53.017    73.400
#> 185663: 959970-99999 HEARD ISLAND (ATLAS COVE)   AS        -53.017    73.400
#> 185664: 959970-99999 HEARD ISLAND (ATLAS COVE)   AS        -53.017    73.400
#>         ELEVATION    BEGIN      END   YEARMODA YEAR MONTH DAY YDAY TEMP
#>      1:         5 19730101 20210116 2010-01-01 2010     1   1    1 29.9
#>      2:         5 19730101 20210116 2010-01-02 2010     1   2    2 29.9
#>      3:         5 19730101 20210116 2010-01-03 2010     1   3    3 29.7
#>      4:         5 19730101 20210116 2010-01-04 2010     1   4    4 29.8
#>      5:         5 19730101 20210116 2010-01-05 2010     1   5    5 29.8
#>     ---                                                                
#> 185660:         4 19980301 20121220 2011-12-27 2011    12  27  361  2.2
#> 185661:         4 19980301 20121220 2011-12-28 2011    12  28  362  2.9
#> 185662:         4 19980301 20121220 2011-12-29 2011    12  29  363  2.4
#> 185663:         4 19980301 20121220 2011-12-30 2011    12  30  364  2.3
#> 185664:         4 19980301 20121220 2011-12-31 2011    12  31  365  2.7
#>         TEMP_ATTRIBUTES DEWP DEWP_ATTRIBUTES    SLP SLP_ATTRIBUTES   STP
#>      1:               8   NA               0     NA              0    NA
#>      2:               9   NA               0     NA              0    NA
#>      3:               9   NA               0     NA              0    NA
#>      4:               7   NA               0     NA              0    NA
#>      5:              12   NA               0     NA              0    NA
#>     ---                                                                 
#> 185660:              24   NA               0  993.1             24 992.7
#> 185661:              24   NA               0  990.7             24 990.3
#> 185662:              24   NA               0  982.7             24 982.3
#> 185663:              23   NA               0 1000.7             23   0.3
#> 185664:              24   NA               0 1012.3             24  11.9
#>         STP_ATTRIBUTES VISIB VISIB_ATTRIBUTES WDSP WDSP_ATTRIBUTES MXSPD GUST
#>      1:              0    NA                0  6.1               8   7.2   NA
#>      2:              0    NA                0  5.6               9   7.2   NA
#>      3:              0    NA                0  7.5               9  10.3   NA
#>      4:              0    NA                0  9.3               7  10.3   NA
#>      5:              0    NA                0  8.8              12  10.8   NA
#>     ---                                                                      
#> 185660:             24    NA                0   NA               0    NA   NA
#> 185661:             24    NA                0   NA               0    NA   NA
#> 185662:             24    NA                0   NA               0    NA   NA
#> 185663:             23    NA                0   NA               0    NA   NA
#> 185664:             24    NA                0   NA               0    NA   NA
#>          MAX MAX_ATTRIBUTES  MIN MIN_ATTRIBUTES PRCP PRCP_ATTRIBUTES SNDP I_FOG
#>      1: 30.3              * 29.1              *    0               I   NA    NA
#>      2: 30.5              * 28.6              *    0               I   NA    NA
#>      3: 30.5              * 29.3              *    0               I   NA    NA
#>      4: 30.0              * 29.5              *    0               I   NA    NA
#>      5: 30.4              * 29.3              *    0               I   NA    NA
#>     ---                                                                        
#> 185660:  4.4              *  0.6              *    0               I   NA    NA
#> 185661:  4.7              *  1.5              *    0               I   NA    NA
#> 185662:  4.6              *  0.3              *    0               I   NA    NA
#> 185663:  3.5              *  0.3              *    0               I   NA    NA
#> 185664:  5.2              *  0.0              *    0               I   NA    NA
#>         I_RAIN_DRIZZLE I_SNOW_ICE I_HAIL I_THUNDER I_TORNADO_FUNNEL EA  ES RH
#>      1:             NA         NA     NA        NA               NA NA 4.2 NA
#>      2:             NA         NA     NA        NA               NA NA 4.2 NA
#>      3:             NA         NA     NA        NA               NA NA 4.2 NA
#>      4:             NA         NA     NA        NA               NA NA 4.2 NA
#>      5:             NA         NA     NA        NA               NA NA 4.2 NA
#>     ---                                                                      
#> 185660:             NA         NA     NA        NA               NA NA 0.7 NA
#> 185661:             NA         NA     NA        NA               NA NA 0.8 NA
#> 185662:             NA         NA     NA        NA               NA NA 0.7 NA
#> 185663:             NA         NA     NA        NA               NA NA 0.7 NA
#> 185664:             NA         NA     NA        NA               NA NA 0.7 NA

# }