Download and Return a Tidy Data Frame of GSOD Weather Data

This function automates downloading, cleaning, reformatting of data from the Global Surface Summary of the Day (GSOD) data provided by the US National Centers for Environmental Information (NCEI), Three additional useful elements: saturation vapour pressure (es), actual vapour pressure (ea) and relative humidity (RH) are calculated and returned in the final data frame.

get_GSOD(years, station = NULL, country = NULL, max_missing = NULL,
  agroclimatology = FALSE)

Arguments

years
Year(s) of weather data to download.

station
Optional. Specify a station or multiple stations for which to retrieve, check and clean weather data using STNID. The NCEI reports years for which the data are available. This function checks against these years. However, not all cases are properly documented and in some cases files may not exist on the FTP server even though it is indicated that data was recorded for the station for a particular year. If a station is specified that does not have an existing file on the server, this function will silently fail and move on to existing files for download and cleaning from the FTP server.

country
Optional. Specify a country for which to retrieve weather data; full name or ISO codes can be used.

max_missing
Optional. The maximum number of days allowed to be missing from a station's data before it is excluded from final file output.

agroclimatology
Optional. Logical. Only clean data for stations between latitudes 60 and -60 for agroclimatology work, defaults to FALSE. Set to TRUE to include only stations within the confines of these latitudes.

Value

A data frame as a tibble object of weather data.

Details

Stations reporting a latitude of < -90 or > 90 or longitude of < -180 or > 180 are removed. Stations may be individually checked for number of missing days to assure data quality and omitting stations with too many missing observations.

All units are converted to International System of Units (SI), e.g. Fahrenheit to Celsius and inches to millimetres.

Alternative elevation measurements are supplied for missing values or values found to be questionable based on the Consultative Group for International Agricultural Research's Consortium for Spatial Information group's (CGIAR-CSI) Shuttle Radar Topography Mission 90 metre (SRTM 90m) digital elevation data based on NASA's original SRTM 90m data. Further information on these data and methods can be found on GSODR's GitHub repository.

Data summarise each year by station, which include vapour pressure and relative humidity elements calculated from existing data in GSOD.

All missing values in resulting files are represented as NA regardless of which field they occur in.

For a complete list of the fields and description of the contents and units, please refer to Appendix 1 in the GSODR vignette, utils::vignette("GSODR", package = "GSODR").

For more information see the description of the data provided by NCEI, http://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt.

Note

While GSODR does not distribute GSOD weather data, users of the data should note the conditions that the U.S. NCEI places upon the GSOD data. “The following data and products may have conditions placed on their international commercial use. They can be used within the U.S. or for non-commercial international activities without restriction. The non-U.S. data cannot be redistributed for commercial purposes. Re-distribution of these data by others must provide this same notification.”

References

Jarvis, A., Reuter, H. I, Nelson, A., Guevara, E. (2008) Hole-filled SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database http://srtm.csi.cgiar.org

See also

Examples

# Download weather station for Toowoomba, Queensland for 2010
t <- get_GSOD(years = 2010, station = "955510-99999")

#> |======================================================|100% ~0 s remaining     

# Download data for Philippines for year 2010 with a maximum of five missing
# days per station allowed.

get_GSOD(years = 2010, country = "Philippines", max_missing = 5)

#> # A tibble: 4,709 x 48
#>    USAF  WBAN  STNID STN_NAME CTRY  STATE CALL    LAT   LON ELEV_M
#>    <chr> <chr> <chr> <chr>    <chr> <chr> <chr> <dbl> <dbl>  <dbl>
#>  1 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  2 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  3 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  4 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  5 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  6 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  7 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  8 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#>  9 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#> 10 9822… 99999 9822… VIGAN    RP    <NA>  RPUQ   17.6  120.    4.9
#> # ... with 4,699 more rows, and 38 more variables: ELEV_M_SRTM_90m <dbl>,
#> #   BEGIN <int>, END <int>, YEARMODA <date>, YEAR <chr>, MONTH <chr>,
#> #   DAY <chr>, YDAY <dbl>, TEMP <dbl>, TEMP_CNT <int>, DEWP <dbl>,
#> #   DEWP_CNT <int>, SLP <dbl>, SLP_CNT <int>, STP <dbl>, STP_CNT <int>,
#> #   VISIB <dbl>, VISIB_CNT <int>, WDSP <dbl>, WDSP_CNT <int>, MXSPD <dbl>,
#> #   GUST <dbl>, MAX <dbl>, MAX_FLAG <chr>, MIN <dbl>, MIN_FLAG <chr>,
#> #   PRCP <dbl>, PRCP_FLAG <chr>, SNDP <dbl>, I_FOG <int>, I_RAIN_DRIZZLE <int>,
#> #   I_SNOW_ICE <int>, I_HAIL <int>, I_THUNDER <int>, I_TORNADO_FUNNEL <int>,
#> #   EA <dbl>, ES <dbl>, RH <dbl>

# Download global GSOD data for agroclimatology work for years 2010 and 2011

get_GSOD(years = 2010:2011, agroclimatology = TRUE)

#> # A tibble: 6,507,032 x 48
#>    USAF  WBAN  STNID STN_NAME CTRY  STATE CALL    LAT   LON ELEV_M
#>    <chr> <chr> <chr> <chr>    <chr> <chr> <chr> <dbl> <dbl>  <dbl>
#>  1 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  2 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  3 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  4 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  5 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  6 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  7 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  8 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#>  9 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#> 10 0082… 99999 0082… WXPOD82… AF    <NA>  <NA>   33.0  65.6  1157.
#> # ... with 6,507,022 more rows, and 38 more variables: ELEV_M_SRTM_90m <dbl>,
#> #   BEGIN <int>, END <int>, YEARMODA <date>, YEAR <chr>, MONTH <chr>,
#> #   DAY <chr>, YDAY <dbl>, TEMP <dbl>, TEMP_CNT <int>, DEWP <dbl>,
#> #   DEWP_CNT <int>, SLP <dbl>, SLP_CNT <int>, STP <dbl>, STP_CNT <int>,
#> #   VISIB <dbl>, VISIB_CNT <int>, WDSP <dbl>, WDSP_CNT <int>, MXSPD <dbl>,
#> #   GUST <dbl>, MAX <dbl>, MAX_FLAG <chr>, MIN <dbl>, MIN_FLAG <chr>,
#> #   PRCP <dbl>, PRCP_FLAG <chr>, SNDP <dbl>, I_FOG <int>, I_RAIN_DRIZZLE <int>,
#> #   I_SNOW_ICE <int>, I_HAIL <int>, I_THUNDER <int>, I_TORNADO_FUNNEL <int>,
#> #   EA <dbl>, ES <dbl>, RH <dbl>