Download and Return a Tidy Data Frame of GSOD Weather Station Data Inventories

The NCEI maintains a document, ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-inventory.txt, which lists the number of weather observations by station-year-month from the beginning of the stations' records. This function retrieves that document and prints an information header displaying the last update time with a data frame of the inventory information for each station-year-month.

get_inventory()

Value

A data frame as a tibble object of station inventories

Note

The GSOD data, which are downloaded and manipulated by GSODR, stipulate that the following notice should be given. “The following data and products may have conditions placed on their international commercial use. They can be used within the U.S. or for non- commercial international activities without restriction. The non-U.S. data cannot be redistributed for commercial purposes. Re-distribution of these data by others must provide this same notification.”

The download process can take quite some time to complete.

Examples

inventory <- get_inventory()
inventory

#>    *** FEDERAL CLIMATE COMPLEX INTEGRATED SURFACE DATA INVENTORY ***   
#>     This inventory provides the number of weather observations by   
#>     STATION-YEAR-MONTH for beginning of record through October 2018    
#> # A tibble: 620,253 x 14
#>    STN…  YEAR   JAN   FEB   MAR   APR   MAY   JUN   JUL   AUG   SEP   OCT   NOV
#>    <ch> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int>
#>  1 007…  2011     0     0  2104  2797  2543  2614   382     0     0     0     0
#>  2 007…  2013     0     0     0     0     0     0   710     0     0     0     0
#>  3 007…  2012     0     0     0     0     0     0   367     0     0     0     0
#>  4 007…  2014     0     0     0     0     0     0   180     0     4     0   552
#>  5 007…  2016     0     0     0     0     0   794     0     0     0     0     0
#>  6 007…  2017     0   914  2626   380   277   406  1230  1010     0     0     0
#>  7 007…  2014     0     0     0     0     0     0     0     0     1     0     0
#>  8 007…  2015     0     0     0     0     0     0     0     0   198     0     0
#>  9 008…  2009  2665  2206  2518  2471  2738  2490  2554  2430  2589  4163  3672
#> 10 008…  2010  4368  3630  3890  3690  3734  1455  1588     0     0     0     0
#> # ... with 620,243 more rows, and 1 more variable: DEC <int>