Download and Return a Tidy Data Frame of GSOD Weather Station Data Inventories

The NCEI maintains a document, ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-inventory.txt, which lists the number of weather observations by station-year-month from the beginning of the stations' records. This function retrieves that document, prints the header to display the last update time and returns a data frame of the inventory information for each station-year-month.

get_inventory()

Value

A data frame as a tibble tibble object of station inventories

Note

The GSOD data, which are downloaded and manipulated by this R package, stipulate that the following notice should be given. “The following data and products may have conditions placed on their international commercial use. They can be used within the U.S. or for non-commercial international activities without restriction. The non-U.S. data cannot be redistributed for commercial purposes. Re-distribution of these data by others must provide this same notification.”

The download process can take quite some time to complete.

Examples

# NOT RUN {
inventory <- get_inventory()
# }